requirements.txt
alembic==1.9.0
amqp==5.1.1
anyio==3.6.2
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
billiard==3.6.4.0
cachelib==0.9.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
comm==0.1.1
config==0.5.1
debugpy==1.6.4
decorator==5.1.1
dnspython==2.2.1
email-validator==1.3.0
entrypoints==0.4
exceptiongroup==1.0.4
executing==1.2.0
Flask==2.2.2
Flask-Login==0.6.2
flask-marshmallow==0.14.0
Flask-Migrate==4.0.0
Flask-Moment==1.0.5
Flask-Session==0.4.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
flatdict==4.0.1
flower==1.2.0
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.2
httpx==0.23.1
humanize==4.4.0
idna==3.4
importlib-metadata==5.1.0
iniconfig==1.1.1
install==1.3.5
ipykernel==6.19.0
ipython==8.7.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonpickle==3.0.1
jupyter_client==7.4.8
jupyter_core==5.1.0
kombu==5.2.4
Levenshtein==0.20.9
Mako==1.2.4
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-sqlalchemy==0.28.1
matplotlib-inline==0.1.6
musicbrainzngs==0.7.1
nest-asyncio==1.5.6
numpy==1.23.5
oauthlib==3.2.2
packaging==22.0
pandas==1.5.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pillowcase==2.0.0
platformdirs==2.6.0
pluggy==1.0.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pylast==5.1.0
pytest==7.2.0
pytest-flask==1.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-Levenshtein==0.20.9
python3-discogs-client==2.5
pytz==2022.6
pyzmq==24.0.1
rapidfuzz==2.13.7
redis==4.4.0
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
scipy==1.9.3
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.44
sqlalchemy-get-or-create==0.1.5
stack-data==0.6.2
thefuzz==0.19.0
tomli==2.0.1
tornado==6.2
traitlets==5.6.0
urllib3==1.26.13
URLObject==2.4.3
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.2.2
WTForms==3.0.1
xmltodict==0.13.0
zipp==3.11.0
